🐛 fix: allow non-admin users to use skill CLI commands
- ListSkills API uses Store.List (DB-filtered) for non-admin users
  instead of requiring admin access
- CLI skill install/remove use profile endpoints instead of admin ones
- Move cliUserSkillsDir to setup_skills.go with its callers
